pyproject.toml
[tool.poetry]
name = "pf2e-tui"
version = "0.1.0"
description = ""
authors = ["JamesTheBard <jamesthebard@gmail.com>"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
textual = "^0.79.1"
rich = "^13.8.1"
pyyaml = "^6.0.2"
python-box = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
